extractor_service/requirements.txt
grpcio
grpcio-tools
pandas
openpyxl
requests
python-dotenv